refactor(server): extract handleUpdatePlayerSkills to dedicated module (S27.8.30)
Polish slice: bring `team-player-handlers.ts` under DoD secondary
400 by extracting `handleUpdatePlayerSkills` (~189 lines) into its
own dedicated module `routes/team-player-skills-handler.ts`
(245 lines, well under 400).

Re-export chain preserved:
- `team-player-skills-handler.ts` exports the handler
- `team-player-handlers.ts` re-exports it
- `team.ts` re-exports from `team-player-handlers.ts`

Public API consumed by `team.test.ts` integration tests stays
unchanged.

Cleanup of orphaned imports in `team-player-handlers.ts` (6
removed): `getNextAdvancementPspCost`,
`getPositionCategoryAccess`, `SKILLS_BY_SLUG`, `SKILLS_DEFINITIONS`,
`AdvancementType`, `PlayerAdvancement`.

`team-player-handlers.ts` shrinks 578 -> 365 lines (**under DoD
secondary 400**).

6 new smoke tests `team-player-skills-handler.test.ts`:
- exports x3 (named + 2 re-exports through chain)
- 400 gates on missing body fields x2
- 404 gate on team not found

- 1842/1842 server tests pass
- typecheck clean
